pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mctoolbox"
version = "0.1.0"
description = "MCAF motor control tuning library using pyX2Cscope"
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = {text = "MIT"}
dependencies = [
    "pyx2cscope>=0.5.0",
    "numpy>=1.26",
    "matplotlib>=3.7",
    "pyyaml>=6.0",
    "control>=0.10",
]

[project.optional-dependencies]
notebook = [
    "jupyterlab>=4.0",
    "ipykernel>=7.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "pyinstaller>=6.0",
]

[project.scripts]
mctoolbox-gui = "mctoolbox.gui:main"

[tool.setuptools.packages.find]
include = ["mctoolbox*"]

[tool.setuptools.package-data]
"mctoolbox.wizards" = ["*.yaml", "*.yml"]
